Completed automatic image search and tweet functionality!
Now 3 times a day, he will look for nature images that other people have tweeted, pixelize the image and tweet it!
