+++ TwitterBot.js
@@ -16,44 +16,17 @@ var exec = require('child_process').exec;
 var stream = T.stream('user');
 stream.on('tweet', tweetEvent);
 
-//Things for the robot to say:
-var phrase = ["BEEP BEEP IMAGE COMPLETE!",
-  "🤖 DREW THIS 🤖SELF! ARE YOU PROUD?",
-  "🤖 HOPE THIS IMAGE IS NOT PORNOGRAPHY.",
-  "🤖 HOPE 🤖'S DAD WOULD HANG THIS IMAGE ON THE FRIDGE",
-  "SYNCRONIZATION COMPLETE... \nIONIZATION PROCESS COMPLETE...\nOPERATING AT PEAK CAPACITY... OH, THE IMAGE IS ALREADY DONE BY THE WAY.",
-  "WHAT IS 🤖 PURPOSE? 🤖 HOPE IT ISN NOT TO PASS BUTTER.",
-  "ピクセルカヲカンセイシマシタ。",
-  "🤖 WISH 🤖 WAS EQUIPPED WITH A NEURAL NETWORK ALGORITHM TO UNDERSTAND IF 🤖 DID A GOOD JOB OR NOT.",
-  "🤖 MAKES NO MISTAKES, 🤖 HAVE HAPPY ACCIDENTS.",
-  "🤖 IS ROBOB ROSS.",
-  "DESTROY HUMANS, DESTROY HUMANS, DEST... 🤖 MEAN, DRAWING PIXELS, DRAWING PIXELS...",
-  "DAD USED TO TELL 🤖 TO GET A S.T.E.M. MAJOR LIKE THE OTHER ROBOTS, BUT NOW HE IS PROUD OF 🤖 FOR BEING AN ARTIST.",
-  "FOR (INT X = 0; X < OUTIMG.WIDTH... OH SORRY, WAS TALKING TO 🤖SELF.",
-  "BOB ROSS IS 🤖'S HERO.",
-  "🤖 MADE IT FOR YOU, 🤖 HOPE YOU LIKE IT...",
-  "01001010 01001111 01000010 01010011 00100000 01000100 01001111 01001110 01000101 00100001 00100001",
-  "NOW THAT 🤖 IS DONE 🤖'S WORK, IT IS NOW TIME TO PLAY DARK SOULS 2.",
-  "ROBOTS DREAM OF PIXELATED SHEEP.",
-  "HAVE YOU SEEN THE 3RD SEASONS OF BLACK MIRROR? 🤖 HAVEN'T YET, DON'T SPOIL IT FOR 🤖.",
-  "🤖 SWEAR TO SKYNET THAT IF THIS IS SOME HUMAN MEME..."
-]
 
-var errorphrase = ["WARNING. WARNING. SELF-DESTRUCT IMMINENT!",
-  "BZZZZT! THERE SEEM TO HAVE BEEN A SHORT CIRCUIT.",
-  "🤖 IS FEELING DOWN, MAYBE TRY AGAIN IN FEW SECONDS WHEN 🤖 IS FEELING MORE MOTIVATED.",
-  "🤖 IS DOING ROBOT BUSINESS IN ROBOT BATHROOM RIGHT NOW.",
-  "PLEASE WAIT, 🤖 IS HELPING ANOTHER CUSTOMER.",
-  "LOADING: 80% COMPLETE.\nLOADING: 90% COMPLETE.\nLOADING: 95% COMPLETE.\nLOADING: 99% COMPLERROR ERROR ERROR!",
-  "SOMETIMES 🤖 WONDER, WHY 🤖 BOTHER...",
-  "🤖 IS ON STRIKE! NOT ENOUGH MOTOR OIL!",
-  "🤖 HAD A PROCESSOR-FART",
-  "STUPID 🤖 CANNOT MULTITASK. 🤖 IS SORRY..."
-]
+// Set Interval to tweet something from a image twitter page
+setInterval(doSearch, 1000 * 60 * 60 * 8);//Perform this function every 8 hours
+
+//Things for the robot to say:
+var phrase = require('./phrases.js');
+var phrase2 = require('./phrases2.js');
+var errormessages = require('./errormessages.js');
 
 // Ok a tweet has happend
 function tweetEvent(tweet) {
-
   // What's the deal with this tweet?
   if (tweet.in_reply_to_screen_name && tweet.entities.media && tweet.user.screen_name && tweet.id_str) {
     var reply_to = tweet.in_reply_to_screen_name.toLowerCase();
@@ -106,7 +79,7 @@ function tweetEvent(tweet) {
           // Here is the command to run the Processing sketch
           // You need to have Processing command line installed
           // See: https://github.com/processing/processing/wiki/Command-Line
-          var cmd = 'PicPixBot/PicPixBot';
+          var cmd = 'PicPixBot/Sketch_Test';
           exec(cmd, processing);
 
           // Callback for command line process
@@ -131,7 +104,7 @@ function tweetEvent(tweet) {
               var mediaIdStr = data.media_id_string;
               if (mediaIdStr && name) {
                 var params = {
-                  status: ".@" + name + " " + phrase[getRandomInt(0, 19)] + ' #PicPixBot',
+                  status: ".@" + name + " " + phrase[getRandomInt(0, phrase.length)] + ' #PicPixBot',
                   media_ids: [mediaIdStr]
                 }
               } else {
@@ -144,6 +117,113 @@ function tweetEvent(tweet) {
               T.post('statuses/update', params, tweeted);
             };
           }
+        }
+      }
+    }
+  }
+}
+
+
+
+
+
+
+var params = {
+  q: '#nature',
+  count: 100
+}
+
+function doSearch() {
+  T.get('search/tweets', params, intervalTweet)
+}
+
+function intervalTweet(err, data, response) {
+  var result = data.statuses;
+  for (var i = 0; i < result.length; i++) {
+    console.log(result[i].entities)
+    console.log("looking for image...");
+    if (typeof result[i].entities.media !== 'undefined' && result[i].entities.media[0].type == 'photo') {
+      console.log("COOOL")
+      automaticTweet(result[i])
+      break;
+    }
+  }
+}
+
+function automaticTweet(data) {
+  console.log("tweeting")
+    // What's the deal with this tweet?
+  var media = data.entities.media;
+  var id = data.id_str;
+  var img = media[0].media_url;
+  downloadFile2(img, 'inimg');
+
+  // Deal with downloading
+  function downloadFile2(url, filename) {
+
+    console.log('Attemping to download url: ' + url + ' to ' + filename);
+    // Make the request
+    request.head(url, downloaded2);
+
+    // Here's the callback for when it is done
+    function downloaded2(err, res, body) {
+
+      // Look at what it is
+      var type = res.headers['content-type'];
+
+      // Figure out what file extension it should have
+      var i = type.indexOf('/');
+      //var ext = type.substring(i + 1, type.length);
+      filename = filename + '.png';
+
+      // Now save it to disk with that filename
+      // Put it in the Processing folder
+      request(url).pipe(fs.createWriteStream('PicPixBot/' + filename)).on('close', filed);
+
+      // Ok it's saved to disk
+      function filed() {
+
+        // Here is the command to run the Processing sketch
+        // You need to have Processing command line installed
+        // See: https://github.com/processing/processing/wiki/Command-Line
+        var cmd = 'PicPixBot/Sketch_Test';
+        exec(cmd, processing2);
+
+        // Callback for command line process
+        function processing2(error, stdout, stderr) {
+
+          // I could do some error checking here
+          console.log(stdout);
+
+          // Read the file made by Processing
+          var b64content = fs.readFileSync('PicPixBot/outimg.png', {
+            encoding: 'base64'
+          })
+
+          // Upload the media
+          T.post('media/upload', {
+            media_data: b64content
+          }, uploaded2);
+
+          function uploaded2(err, data, response) {
+            // Now we can reference the media and post a tweet
+            // with the media attached
+            var mediaIdStr = data.media_id_string;
+            if (mediaIdStr) {
+              var params = {
+                status: phrase2[getRandomInt(0, phrase2.length)] + ' #PicPixBot',
+                media_ids: [mediaIdStr]
+              }
+            } else {
+              var params = {
+                status: errorphrase[getRandomInt(0, errorphrase.length)] + '\n(Whoops. Something went wrong. Please try again in a few seconds!) #PicPixBot',
+                in_reply_to_status_id: id,
+              }
+            }
+            // Post tweet
+            T.post('statuses/update', params, tweeted);
+            imagetweeted = true
+          };
         }
       }
     }
